refactor(npm): create general use `NpmPackageResolver`
